+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/windows/WndWandmaker.java
@@ -3,7 +3,7 @@
  * Copyright (C) 2012-2015 Oleg Dolya
  *
  * Shattered Pixel Dungeon
- * Copyright (C) 2014-2019 Evan Debenham
+ * Copyright (C) 2014-2023 Evan Debenham
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -26,11 +26,12 @@ import com.shatteredpixel.shatteredpixeldungeon.actors.mobs.npcs.Wandmaker;
 import com.shatteredpixel.shatteredpixeldungeon.items.Item;
 import com.shatteredpixel.shatteredpixeldungeon.items.quest.CorpseDust;
 import com.shatteredpixel.shatteredpixeldungeon.items.quest.Embers;
-import com.shatteredpixel.shatteredpixeldungeon.items.wands.Wand;
 import com.shatteredpixel.shatteredpixeldungeon.messages.Messages;
 import com.shatteredpixel.shatteredpixeldungeon.plants.Rotberry;
+import com.shatteredpixel.shatteredpixeldungeon.scenes.GameScene;
 import com.shatteredpixel.shatteredpixeldungeon.scenes.PixelScene;
 import com.shatteredpixel.shatteredpixeldungeon.sprites.ItemSprite;
+import com.shatteredpixel.shatteredpixeldungeon.ui.ItemButton;
 import com.shatteredpixel.shatteredpixeldungeon.ui.RedButton;
 import com.shatteredpixel.shatteredpixeldungeon.ui.RenderedTextBlock;
 import com.shatteredpixel.shatteredpixeldungeon.ui.Window;
@@ -39,12 +40,19 @@ import com.shatteredpixel.shatteredpixeldungeon.utils.GLog;
 public class WndWandmaker extends Window {
 
 	private static final int WIDTH		= 120;
-	private static final int BTN_HEIGHT	= 20;
-	private static final float GAP		= 2;
-	
+	private static final int BTN_SIZE	= 32;
+	private static final int BTN_GAP	= 5;
+	private static final int GAP		= 2;
+
+	Wandmaker wandmaker;
+	Item questItem;
+
 	public WndWandmaker( final Wandmaker wandmaker, final Item item ) {
 		
 		super();
+
+		this.wandmaker = wandmaker;
+		this.questItem = item;
 		
 		IconTitle titlebar = new IconTitle();
 		titlebar.icon(new ItemSprite(item.image(), null));
@@ -65,46 +73,90 @@ public class WndWandmaker extends Window {
 		message.maxWidth(WIDTH);
 		message.setPos(0, titlebar.bottom() + GAP);
 		add( message );
-		
-		RedButton btnWand1 = new RedButton( Messages.titleCase(Wandmaker.Quest.wand1.name()) ) {
+
+		ItemButton btnWand1 = new ItemButton(){
 			@Override
 			protected void onClick() {
-				selectReward( wandmaker, item, Wandmaker.Quest.wand1 );
+				if (Dungeon.hero.belongings.contains(questItem)) {
+					GameScene.show(new RewardWindow(item()));
+				} else {
+					hide();
+				}
 			}
 		};
-		btnWand1.setRect(0, message.top() + message.height() + GAP, WIDTH, BTN_HEIGHT);
+		btnWand1.item(Wandmaker.Quest.wand1);
+		btnWand1.setRect( (WIDTH - BTN_GAP) / 2 - BTN_SIZE, message.top() + message.height() + BTN_GAP, BTN_SIZE, BTN_SIZE );
 		add( btnWand1 );
 		
-		RedButton btnWand2 = new RedButton( Messages.titleCase(Wandmaker.Quest.wand2.name()) ) {
+		ItemButton btnWand2 = new ItemButton(){
 			@Override
 			protected void onClick() {
-				selectReward( wandmaker, item, Wandmaker.Quest.wand2 );
+				if (Dungeon.hero.belongings.contains(questItem)) {
+					GameScene.show(new RewardWindow(item()));
+				} else {
+					hide();
+				}
 			}
 		};
-		btnWand2.setRect(0, btnWand1.bottom() + GAP, WIDTH, BTN_HEIGHT);
-		add( btnWand2 );
+		btnWand2.item(Wandmaker.Quest.wand2);
+		btnWand2.setRect( btnWand1.right() + BTN_GAP, btnWand1.top(), BTN_SIZE, BTN_SIZE );
+		add(btnWand2);
 		
 		resize(WIDTH, (int) btnWand2.bottom());
 	}
 	
-	private void selectReward( Wandmaker wandmaker, Item item, Wand reward ) {
-		
-		hide();
-		
-		item.detach( Dungeon.hero.belongings.backpack );
+	private void selectReward( Item reward ) {
 
-		reward.identify();
+		if (reward == null){
+			return;
+		}
+
+		hide();
+
+		questItem.detach( Dungeon.hero.belongings.backpack );
+
+		reward.identify(false);
 		if (reward.doPickUp( Dungeon.hero )) {
-			GLog.i( Messages.get(Dungeon.hero, "you_now_have", reward.name()) );
+			GLog.i( Messages.capitalize(Messages.get(Dungeon.hero, "you_now_have", reward.name())) );
 		} else {
 			Dungeon.level.drop( reward, wandmaker.pos ).sprite.drop();
 		}
 		
-		wandmaker.yell( Messages.get(this, "farewell", Dungeon.hero.name()) );
+		wandmaker.yell( Messages.get(this, "farewell", Messages.titleCase(Dungeon.hero.name())) );
 		wandmaker.destroy();
 		
 		wandmaker.sprite.die();
 		
 		Wandmaker.Quest.complete();
 	}
+
+	private class RewardWindow extends WndInfoItem {
+
+		public RewardWindow( Item item ) {
+			super(item);
+
+			RedButton btnConfirm = new RedButton(Messages.get(WndSadGhost.class, "confirm")){
+				@Override
+				protected void onClick() {
+					RewardWindow.this.hide();
+
+					selectReward( item );
+				}
+			};
+			btnConfirm.setRect(0, height+2, width/2-1, 16);
+			add(btnConfirm);
+
+			RedButton btnCancel = new RedButton(Messages.get(WndSadGhost.class, "cancel")){
+				@Override
+				protected void onClick() {
+					hide();
+				}
+			};
+			btnCancel.setRect(btnConfirm.right()+2, height+2, btnConfirm.width(), 16);
+			add(btnCancel);
+
+			resize(width, (int)btnCancel.bottom());
+		}
+	}
+
 }
